Bootstrap3
Added NuGet package restore functionality. Added NuGet reference to MVC
4 for ease of development, but .gitignore ensures all NuGet packages are
not included in the branch.

Included a .nuspec file to improve the NuGet experience. When
downloading the MvcPaging package, one might not realize they need the
Microsoft Unobtrusive Ajax library, etc. The nuspec file will make sure
that the user downloading MvcPaging will have everything they need for
it to work.

To work with Bootstrap 3, see changes to AjaxPager.cs. Bootstrap 3
changes how the pagination container works, so instead of putting
class="pagination..." on the containing div, the ul inside of it now has
class="pagination..." (this is in reference to the page numbers
previous/next buttons below the list of paged items). To compensate for
this change, a new class name, pagination-container, was given to the
containing div so that customization, such as centered or right aligned,
can still be made.